requirements.txt
torch==2.0.1
torchvision==0.15.2
efficientnet_pytorch==0.7.1
pytorch_lightning==2.0.9
transformers==4.31.0
opencv-python-headless==4.8.0.76
celery[redis]==5.3.4
fpdf2==2.7.7
wordcloud==1.9.2
matplotlib==3.7.2
pandas==2.0.3
scikit-learn==1.3.0
uvicorn==0.22.0
gunicorn==20.1.0
torchmetrics==1.0.3
nvidia-pyindex
nvidia-tensorrt